[feature-android] : now stompClient can handling result with consumer callback
